src/requirements.txt
neo4j
python-dotenv
passlib
pydantic
fastapi[all]
python-jose[cryptography]
uvicorn[standard]
passlib[bcrypt]
pytest
pytest-asyncio
pytest-order
starlette
